refactor: 계정 생성 시 기본 skinId 지정

+++ src/main/java/com/be/dohands/member/dto/MemberResponse.java
@@ -21,7 +21,7 @@ public class MemberResponse {
     private LocalDate hireDate;
     private String jobGroup;
     private String jobCategory;
-    private Long skinId;
+    private Integer skinId;
     private Role role;
 
     public MemberResponse(Member member, String level) {
